+++ packages/wallet/frontend/src/components/providers/index.tsx
@@ -4,7 +4,6 @@ import { OnboardingProvider } from './OnboardingProvider'
 import { PasswordProvider } from './PasswordProvider'
 import { ThemeProvider } from 'next-themes'
 import { MenuProvider } from './MenuProvider'
-import { CardProvider } from './CardProvider'
 
 type AppProviderProps = {
   children: ReactNode
@@ -21,9 +20,7 @@ export const AppProvider = ({ children }: AppProviderProps) => {
       <MenuProvider>
         <OnboardingProvider>
           <PasswordProvider>
-            <CardProvider>
-              <DialogProvider>{children}</DialogProvider>
-            </CardProvider>
+            <DialogProvider>{children}</DialogProvider>
           </PasswordProvider>
         </OnboardingProvider>
       </MenuProvider>
