+++ packages/wallet/frontend/src/pages/card/index.tsx
@@ -2,20 +2,33 @@ import { AppLayout } from '@/components/layouts/AppLayout'
 import { PageHeader } from '@/components/PageHeader'
 import { CardButtonMenu } from '@/components/userCards/CardButtonMenu'
 import { UserCard } from '@/components/userCards/UserCard'
-import { useCardContext } from '@/lib/context/card'
 import { NextPageWithLayout } from '@/lib/types/app'
+import { useState } from 'react'
 
+const CARD_TYPES = {
+  normal: 'normal',
+  details: 'details',
+  frozen: 'frozen'
+} as const
+
+export type CardTypes = keyof typeof CARD_TYPES
+
+function CardContainer() {
+  const [state, setState] = useState<CardTypes>('normal')
+  return (
+    <div className="flex flex-col gap-6 max-w-[329px] justify-center items-center">
+      <UserCard type={state} />
+      <CardButtonMenu fn={setState} />
+    </div>
+  )
+}
 const UserCardPage: NextPageWithLayout = () => {
-  const { cardType } = useCardContext()
   return (
     <>
       <div className="flex items-center justify-between md:flex-col md:items-start md:justify-start">
         <PageHeader title="Your Card" />
       </div>
-      <div className="flex justify-center items-center flex-col gap-6">
-        <UserCard type={cardType} />
-        <CardButtonMenu />
-      </div>
+      <CardContainer />
     </>
   )
 }
